Plant stuff

+++ src/ecologysim/IPlantDB.java
@@ -1,6 +1,6 @@
 package ecologysim;
 
 public interface IPlantDB {
-	void addPlant(IPlantSpecies species, int[] produceMasses, int multiplier);
-	void addPlant(IPlantSpecies species, int count, int[] produceMasses, int multiplier);
+	void addPlant(IPlantSpecies species, double[] produceMasses, int multiplier);
+	void addPlant(IPlantSpecies species, int count, double[] produceMasses, int multiplier);
 }
